Fixed issue with bad tar files.
When extracting tar files, the system tar function is used first. If it
fails, the Python tar function is used. A backup copy of the tar file is
made when issues are detected with the tar file. This backup must be
used if Python extraction fails. If Python extraction succeeds, the
backup file is removed.